Move drawable timing changes to Timing/Drawables.
