+++ osu.Game.Rulesets.Mania/Timing/TimingChangeContainer.cs
@@ -10,6 +10,7 @@ using OpenTK;
 using osu.Game.Beatmaps.ControlPoints;
 using osu.Game.Rulesets.Objects;
 using osu.Game.Rulesets.Objects.Drawables;
+using osu.Game.Rulesets.Mania.Timing.Drawables;
 
 namespace osu.Game.Rulesets.Mania.Timing
 {
