+++ osu.Game.Rulesets.Mania/UI/ManiaHitRenderer.cs
@@ -21,7 +21,7 @@ using osu.Game.Rulesets.Mania.Objects;
 using osu.Game.Rulesets.Mania.Objects.Drawables;
 using osu.Game.Rulesets.Mania.Scoring;
 using osu.Game.Rulesets.Mania.Timing;
-using osu.Game.Rulesets.Mania.Timing.Drawable;
+using osu.Game.Rulesets.Mania.Timing.Drawables;
 using osu.Game.Rulesets.Objects.Drawables;
 using osu.Game.Rulesets.Objects.Types;
 using osu.Game.Rulesets.Scoring;
