+++ osu.Game.Rulesets.Mania/osu.Game.Rulesets.Mania.csproj
@@ -83,7 +83,9 @@
     <Compile Include="ManiaRuleset.cs" />
     <Compile Include="Mods\ManiaMod.cs" />
     <Compile Include="UI\SpecialColumnPosition.cs" />
-    <Compile Include="Timing\DrawableTimingChange.cs" />
+    <Compile Include="Timing\Drawables\DrawableGravityTimingChange.cs" />
+    <Compile Include="Timing\Drawables\DrawableScrollingTimingChange.cs" />
+    <Compile Include="Timing\Drawables\DrawableTimingChange.cs" />
     <Compile Include="Timing\TimingChange.cs" />
     <Compile Include="Timing\TimingChangeContainer.cs" />
   </ItemGroup>
